Use exec to gather bind credentials from the apb
bind-init: The bind-init script will be called after a bind
call is made.  The goal of this script it to prevent the container
from exiting until the broker has had an oportunity to gather
the credentials.

broker-bind-creds: The broker-bind-creds script will be called by
the broker when it execs into the container. The script is going
to wait until /etc/asb/bind-creds has been filled with the bind
credentials. Once the bind credentials are recieved, broker-bind-creds
will delete /etc/asb/bind-creds which will signal bind-init to exit
allowing the container to exit.

apb bind: The expectation for the user creating an apb is to gather
bind credentials and write them to the file '/etc/apb/bind-creds'.
After that, the broker will take care of the rest.